refactor: Split in ItemHierarchyPublisher and ContainerPublisher
