+++ src/library-authoring/containers/messages.ts
@@ -51,86 +51,6 @@ const messages = defineMessages({
     defaultMessage: 'Content Hierarchy',
     description: 'Heading for usage tab hierarchy section',
   },
-  publishContainerConfirmHeading: {
-    id: 'course-authoring.library-authoring.container-sidebar.publish-confirm-heading',
-    defaultMessage: 'Confirm Publish',
-    description: 'Header text shown while confirming publish of a unit/subsection/section',
-  },
-  publishContainerConfirm: {
-    id: 'course-authoring.library-authoring.container-sidebar.publish-confirm-button',
-    defaultMessage: 'Publish',
-    description: 'Button text shown to confirm publish of a unit/subsection/section',
-  },
-  publishContainerCancel: {
-    id: 'course-authoring.library-authoring.container-sidebar.publish-cancel',
-    defaultMessage: 'Cancel',
-    description: 'Button text shown to cancel publish of a unit/subsection/section',
-  },
-  publishContainerSuccess: {
-    id: 'course-authoring.library-authoring.container-sidebar.publish-success',
-    defaultMessage: 'All changes published',
-    description: 'Popup text after publishing a unit/subsection/section',
-  },
-  publishContainerFailed: {
-    id: 'course-authoring.library-authoring.container-sidebar.publish-failure',
-    defaultMessage: 'Failed to publish changes',
-    description: 'Popup text seen if publishing a unit/subsection/section fails',
-  },
-  publishSectionWarning: {
-    id: 'course-authoring.library-authoring.section-sidebar.publish-empty-warning',
-    defaultMessage: 'This section will be <highlight>published</highlight>.',
-    description: 'Content details shown before publishing an empty section',
-  },
-  publishSectionWithChildrenWarning: {
-    id: 'course-authoring.library-authoring.section-sidebar.publish-warning',
-    defaultMessage: (
-      'This section and the {childCount, plural, one {subsection} other {subsections}}'
-      + ' it contains will all be <highlight>published</highlight>.'
-    ),
-    description: 'Content details shown before publishing a section that contains subsections',
-  },
-  publishSubsectionWarning: {
-    id: 'course-authoring.library-authoring.subsection-sidebar.publish-empty-warning',
-    defaultMessage: 'This subsection will be <highlight>published</highlight>.',
-    description: 'Content details shown before publishing an empty subsection',
-  },
-  publishSubsectionWithChildrenWarning: {
-    id: 'course-authoring.library-authoring.subsection-sidebar.publish-warning',
-    defaultMessage: (
-      'This subsection and the {childCount, plural, one {unit} other {units}}'
-      + ' it contains will all be <highlight>published</highlight>.'
-    ),
-    description: 'Content details shown before publishing a subsection that contains units',
-  },
-  publishSubsectionWithParentWarning: {
-    id: 'course-authoring.library-authoring.subsection-sidebar.publish-parent-warning',
-    defaultMessage: (
-      'Its {parentCount, plural, one {parent section} other {parent sections}}'
-      + ' will be <highlight>draft</highlight>.'
-    ),
-    description: 'Parent details shown before publishing a unit that has one or more parent subsections',
-  },
-  publishUnitWarning: {
-    id: 'course-authoring.library-authoring.unit-sidebar.publish-empty-warning',
-    defaultMessage: 'This unit will be <highlight>published</highlight>.',
-    description: 'Content details shown before publishing an empty unit',
-  },
-  publishUnitWithChildrenWarning: {
-    id: 'course-authoring.library-authoring.unit-sidebar.publish-warning',
-    defaultMessage: (
-      'This unit and the {childCount, plural, one {component} other {components}}'
-      + ' it contains will all be <highlight>published</highlight>.'
-    ),
-    description: 'Content details shown before publishing a unit that contains components',
-  },
-  publishUnitWithParentWarning: {
-    id: 'course-authoring.library-authoring.unit-sidebar.publish-parent-warning',
-    defaultMessage: (
-      'Its {parentCount, plural, one {parent subsection} other {parent subsections}}'
-      + ' will be <highlight>draft</highlight>.'
-    ),
-    description: 'Parent details shown before publishing a unit that has one or more parent subsections',
-  },
   settingsTabTitle: {
     id: 'course-authoring.library-authoring.container-sidebar.settings-tab.title',
     defaultMessage: 'Settings',
